Cardano.Ledger.Allegra.Translation

Documentation

shelleyToAllegraAVVMsToDelete ∷ NewEpochState ShelleyEra → UTxO ShelleyEra

Return the subset of UTxO corresponding to Byron-era AVVM addresses, which are to be removed on the Shelley/Allegra boundary. This set will be passed _back_ to the translation functions as the UTxO, allowing these addresses to be removed. This is needed because we cannot do a full scan on the UTxO at this point, since it has been persisted to disk.

Orphan instances

TranslateEra AllegraEra Tx

Instance details

Associated Types

type TranslationError AllegraEra Tx

Methods

translateEra ∷ TranslationContext AllegraEra → Tx (PreviousEra AllegraEra) → Except (TranslationError AllegraEra Tx) (Tx AllegraEra)

TranslateEra AllegraEra PParams

Instance details

Associated Types

type TranslationError AllegraEra PParams

Methods

translateEra ∷ TranslationContext AllegraEra → PParams (PreviousEra AllegraEra) → Except (TranslationError AllegraEra PParams) (PParams AllegraEra)

TranslateEra AllegraEra PParamsUpdate

Instance details

Associated Types

type TranslationError AllegraEra PParamsUpdate

Methods

translateEra ∷ TranslationContext AllegraEra → PParamsUpdate (PreviousEra AllegraEra) → Except (TranslationError AllegraEra PParamsUpdate) (PParamsUpdate AllegraEra)

TranslateEra AllegraEra CommitteeState

Instance details

Associated Types

type TranslationError AllegraEra CommitteeState

Methods

translateEra ∷ TranslationContext AllegraEra → CommitteeState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra CommitteeState) (CommitteeState AllegraEra)

TranslateEra AllegraEra DState

Instance details

Associated Types

type TranslationError AllegraEra DState

Methods

translateEra ∷ TranslationContext AllegraEra → DState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra DState) (DState AllegraEra)

TranslateEra AllegraEra PState

Instance details

Associated Types

type TranslationError AllegraEra PState

Methods

translateEra ∷ TranslationContext AllegraEra → PState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra PState) (PState AllegraEra)

TranslateEra AllegraEra FuturePParams

Instance details

Associated Types

type TranslationError AllegraEra FuturePParams

Methods

translateEra ∷ TranslationContext AllegraEra → FuturePParams (PreviousEra AllegraEra) → Except (TranslationError AllegraEra FuturePParams) (FuturePParams AllegraEra)

TranslateEra AllegraEra UTxO

Instance details

Associated Types

type TranslationError AllegraEra UTxO

Methods

translateEra ∷ TranslationContext AllegraEra → UTxO (PreviousEra AllegraEra) → Except (TranslationError AllegraEra UTxO) (UTxO AllegraEra)

TranslateEra AllegraEra ShelleyGovState

Instance details

Associated Types

type TranslationError AllegraEra ShelleyGovState

Methods

translateEra ∷ TranslationContext AllegraEra → ShelleyGovState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra ShelleyGovState) (ShelleyGovState AllegraEra)

TranslateEra AllegraEra EpochState

Instance details

Associated Types

type TranslationError AllegraEra EpochState

Methods

translateEra ∷ TranslationContext AllegraEra → EpochState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra EpochState) (EpochState AllegraEra)

TranslateEra AllegraEra LedgerState

Instance details

Associated Types

type TranslationError AllegraEra LedgerState

Methods

translateEra ∷ TranslationContext AllegraEra → LedgerState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra LedgerState) (LedgerState AllegraEra)

TranslateEra AllegraEra NewEpochState

Instance details

Associated Types

type TranslationError AllegraEra NewEpochState

Methods

translateEra ∷ TranslationContext AllegraEra → NewEpochState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra NewEpochState) (NewEpochState AllegraEra)

TranslateEra AllegraEra UTxOState

Instance details

Associated Types

type TranslationError AllegraEra UTxOState

Methods

translateEra ∷ TranslationContext AllegraEra → UTxOState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra UTxOState) (UTxOState AllegraEra)

TranslateEra AllegraEra ProposedPPUpdates

Instance details

Associated Types

type TranslationError AllegraEra ProposedPPUpdates

Methods

translateEra ∷ TranslationContext AllegraEra → ProposedPPUpdates (PreviousEra AllegraEra) → Except (TranslationError AllegraEra ProposedPPUpdates) (ProposedPPUpdates AllegraEra)

TranslateEra AllegraEra Update

Instance details

Associated Types

type TranslationError AllegraEra Update

Methods

translateEra ∷ TranslationContext AllegraEra → Update (PreviousEra AllegraEra) → Except (TranslationError AllegraEra Update) (Update AllegraEra)

TranslateEra AllegraEra ShelleyAccounts

Instance details

Associated Types

type TranslationError AllegraEra ShelleyAccounts

Methods

translateEra ∷ TranslationContext AllegraEra → ShelleyAccounts (PreviousEra AllegraEra) → Except (TranslationError AllegraEra ShelleyAccounts) (ShelleyAccounts AllegraEra)

TranslateEra AllegraEra ShelleyCertState

Instance details

Associated Types

type TranslationError AllegraEra ShelleyCertState

Methods

translateEra ∷ TranslationContext AllegraEra → ShelleyCertState (PreviousEra AllegraEra) → Except (TranslationError AllegraEra ShelleyCertState) (ShelleyCertState AllegraEra)

TranslateEra AllegraEra ShelleyInstantStake

Instance details

Associated Types

type TranslationError AllegraEra ShelleyInstantStake

Methods

translateEra ∷ TranslationContext AllegraEra → ShelleyInstantStake (PreviousEra AllegraEra) → Except (TranslationError AllegraEra ShelleyInstantStake) (ShelleyInstantStake AllegraEra)

TranslateEra AllegraEra ShelleyTxOut

Instance details

Associated Types

type TranslationError AllegraEra ShelleyTxOut

Methods

translateEra ∷ TranslationContext AllegraEra → ShelleyTxOut (PreviousEra AllegraEra) → Except (TranslationError AllegraEra ShelleyTxOut) (ShelleyTxOut AllegraEra)

TranslateEra AllegraEra ShelleyTxWits

Instance details

Associated Types

type TranslationError AllegraEra ShelleyTxWits

Methods

translateEra ∷ TranslationContext AllegraEra → ShelleyTxWits (PreviousEra AllegraEra) → Except (TranslationError AllegraEra ShelleyTxWits) (ShelleyTxWits AllegraEra)