Cardano.Ledger.Allegra.Scripts

Documentation

class ShelleyEraScript era ⇒ AllegraEraScript era where

Methods

mkTimeStart ∷ SlotNo → NativeScript era

getTimeStart ∷ NativeScript era → Maybe SlotNo

mkTimeExpire ∷ SlotNo → NativeScript era

getTimeExpire ∷ NativeScript era → Maybe SlotNo

mkRequireSignatureTimelock ∷ ∀ era. Era era ⇒ KeyHash 'Witness → Timelock era

getRequireSignatureTimelock ∷ Timelock era → Maybe (KeyHash 'Witness)

mkRequireAllOfTimelock ∷ ∀ era. Era era ⇒ StrictSeq (Timelock era) → Timelock era

getRequireAllOfTimelock ∷ Timelock era → Maybe (StrictSeq (Timelock era))

mkRequireAnyOfTimelock ∷ ∀ era. Era era ⇒ StrictSeq (Timelock era) → Timelock era

getRequireAnyOfTimelock ∷ Timelock era → Maybe (StrictSeq (Timelock era))

mkRequireMOfTimelock ∷ ∀ era. Era era ⇒ Int → StrictSeq (Timelock era) → Timelock era

getRequireMOfTimelock ∷ Timelock era → Maybe (Int, StrictSeq (Timelock era))

mkTimeStartTimelock ∷ ∀ era. Era era ⇒ SlotNo → Timelock era

getTimeStartTimelock ∷ Timelock era → Maybe SlotNo

mkTimeExpireTimelock ∷ ∀ era. Era era ⇒ SlotNo → Timelock era

getTimeExpireTimelock ∷ Timelock era → Maybe SlotNo

newtype Timelock era

Constructors

MkTimelock (MemoBytes (TimelockRaw era))

Bundled Patterns

pattern TimelockConstr ∷ MemoBytes (TimelockRaw era) → Timelock era

Deprecated: In favor of more consistently name MkTimelock

pattern RequireTimeExpire ∷ AllegraEraScript era ⇒ SlotNo → NativeScript era

pattern RequireTimeStart ∷ AllegraEraScript era ⇒ SlotNo → NativeScript era

data TimelockRaw era

Constructors

TimelockSignature !(KeyHash 'Witness)

TimelockAllOf !(StrictSeq (Timelock era))

TimelockAnyOf !(StrictSeq (Timelock era))

TimelockMOf !Int !(StrictSeq (Timelock era))

TimelockTimeStart !SlotNo

TimelockTimeExpire !SlotNo

inInterval ∷ SlotNo → ValidityInterval → Bool

Test if a slot is in the Validity interval. Recall that a ValidityInterval is a half Open interval, that is why we use (slot < top)

showTimelock ∷ AllegraEraScript era ⇒ NativeScript era → String

evalTimelock ∷ (AllegraEraScript era, NativeScript era ~ Timelock era) ⇒ Set (KeyHash 'Witness) → ValidityInterval → NativeScript era → Bool

eqTimelockRaw ∷ Timelock era → Timelock era → Bool

Check the equality of two underlying types, while ignoring their binary representation, which Eq instance normally does. This is used for testing.

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore ∷ !(StrictMaybe SlotNo)

invalidHereafter ∷ !(StrictMaybe SlotNo)

encodeVI ∷ ValidityInterval → Encode ('Closed 'Dense) ValidityInterval

decodeVI ∷ Decode ('Closed 'Dense) ValidityInterval

translateTimelock ∷ ∀ era1 era2. (Era era1, Era era2) ⇒ Timelock era1 → Timelock era2

This function deconstructs and then reconstructs the timelock script to prove the compiler that we can arbirarily switch out the eras as long as the cryptos for both eras are the same.

upgradeMultiSig ∷ NativeScript ShelleyEra → NativeScript AllegraEra

lteNegInfty ∷ SlotNo → StrictMaybe SlotNo → Bool

less-than-equal comparison, where Nothing is negative infinity

ltePosInfty ∷ StrictMaybe SlotNo → SlotNo → Bool

less-than-equal comparison, where Nothing is positive infinity

Orphan instances

EraScript AllegraEra
Since Timelock scripts are a strictly backwards compatible extension of MultiSig scripts, we can use the same scriptPrefixTag tag here as we did for the ValidateScript instance in MultiSig

Instance details

Associated Types

type Script AllegraEra = (r ∷ Type)

type NativeScript AllegraEra = (r ∷ Type)

Methods

upgradeScript ∷ Script (PreviousEra AllegraEra) → Script AllegraEra

scriptPrefixTag ∷ Script AllegraEra → ByteString

getNativeScript ∷ Script AllegraEra → Maybe (NativeScript AllegraEra)

fromNativeScript ∷ NativeScript AllegraEra → Script AllegraEra

ShelleyEraScript AllegraEra

Instance details

Methods

mkRequireSignature ∷ KeyHash 'Witness → NativeScript AllegraEra

getRequireSignature ∷ NativeScript AllegraEra → Maybe (KeyHash 'Witness)

mkRequireAllOf ∷ StrictSeq (NativeScript AllegraEra) → NativeScript AllegraEra

getRequireAllOf ∷ NativeScript AllegraEra → Maybe (StrictSeq (NativeScript AllegraEra))

mkRequireAnyOf ∷ StrictSeq (NativeScript AllegraEra) → NativeScript AllegraEra

getRequireAnyOf ∷ NativeScript AllegraEra → Maybe (StrictSeq (NativeScript AllegraEra))

mkRequireMOf ∷ Int → StrictSeq (NativeScript AllegraEra) → NativeScript AllegraEra

getRequireMOf ∷ NativeScript AllegraEra → Maybe (Int, StrictSeq (NativeScript AllegraEra))