Cardano.Ledger.Allegra.Scripts

Documentation

class ShelleyEraScript era ⇒ AllegraEraScript era where

Methods

mkTimeStart ∷ SlotNo → NativeScript era

getTimeStart ∷ NativeScript era → Maybe SlotNo

mkTimeExpire ∷ SlotNo → NativeScript era

getTimeExpire ∷ NativeScript era → Maybe SlotNo

mkRequireSignatureTimelock ∷ Era era ⇒ KeyHash 'Witness (EraCrypto era) → Timelock era

getRequireSignatureTimelock ∷ Era era ⇒ Timelock era → Maybe (KeyHash 'Witness (EraCrypto era))

mkRequireAllOfTimelock ∷ Era era ⇒ StrictSeq (Timelock era) → Timelock era

getRequireAllOfTimelock ∷ Era era ⇒ Timelock era → Maybe (StrictSeq (Timelock era))

mkRequireAnyOfTimelock ∷ Era era ⇒ StrictSeq (Timelock era) → Timelock era

getRequireAnyOfTimelock ∷ Era era ⇒ Timelock era → Maybe (StrictSeq (Timelock era))

mkRequireMOfTimelock ∷ Era era ⇒ Int → StrictSeq (Timelock era) → Timelock era

getRequireMOfTimelock ∷ Era era ⇒ Timelock era → Maybe (Int, StrictSeq (Timelock era))

mkTimeStartTimelock ∷ Era era ⇒ SlotNo → Timelock era

getTimeStartTimelock ∷ Era era ⇒ Timelock era → Maybe SlotNo

mkTimeExpireTimelock ∷ Era era ⇒ SlotNo → Timelock era

getTimeExpireTimelock ∷ Era era ⇒ Timelock era → Maybe SlotNo

data Timelock era

pattern RequireTimeExpire ∷ AllegraEraScript era ⇒ SlotNo → NativeScript era

pattern RequireTimeStart ∷ AllegraEraScript era ⇒ SlotNo → NativeScript era

data TimelockRaw era

pattern TimelockConstr ∷ MemoBytes TimelockRaw era → Timelock era

inInterval ∷ SlotNo → ValidityInterval → Bool

Test if a slot is in the Validity interval. Recall that a ValidityInterval is a half Open interval, that is why we use (slot < top)

showTimelock ∷ AllegraEraScript era ⇒ NativeScript era → String

evalTimelock ∷ AllegraEraScript era ⇒ Set (KeyHash 'Witness (EraCrypto era)) → ValidityInterval → NativeScript era → Bool

eqTimelockRaw ∷ Timelock era → Timelock era → Bool

Check the equality of two underlying types, while ignoring their binary representation, which Eq instance normally does. This is used for testing.

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore ∷ !(StrictMaybe SlotNo)

invalidHereafter ∷ !(StrictMaybe SlotNo)

encodeVI ∷ ValidityInterval → Encode ('Closed 'Dense) ValidityInterval

decodeVI ∷ Decode ('Closed 'Dense) ValidityInterval

translateTimelock ∷ ∀ era1 era2. (Era era1, Era era2, EraCrypto era1 ~ EraCrypto era2) ⇒ Timelock era1 → Timelock era2

This function deconstructs and then reconstructs the timelock script to prove the compiler that we can arbirarily switch out the eras as long as the cryptos for both eras are the same.

Orphan instances

Crypto c ⇒ EraScript (AllegraEra c)
Since Timelock scripts are a strictly backwards compatible extension of MultiSig scripts, we can use the same scriptPrefixTag tag here as we did for the ValidateScript instance in MultiSig

Instance details

Associated Types

type Script (AllegraEra c) = (r ∷ Type)

type NativeScript (AllegraEra c) = (r ∷ Type)

Methods

upgradeScript ∷ Script (PreviousEra (AllegraEra c)) → Script (AllegraEra c)

scriptPrefixTag ∷ Script (AllegraEra c) → ByteString

getNativeScript ∷ Script (AllegraEra c) → Maybe (NativeScript (AllegraEra c))

fromNativeScript ∷ NativeScript (AllegraEra c) → Script (AllegraEra c)

Crypto c ⇒ ShelleyEraScript (AllegraEra c)

Instance details

Methods

mkRequireSignature ∷ KeyHash 'Witness (EraCrypto (AllegraEra c)) → NativeScript (AllegraEra c)

getRequireSignature ∷ NativeScript (AllegraEra c) → Maybe (KeyHash 'Witness (EraCrypto (AllegraEra c)))

mkRequireAllOf ∷ StrictSeq (NativeScript (AllegraEra c)) → NativeScript (AllegraEra c)

getRequireAllOf ∷ NativeScript (AllegraEra c) → Maybe (StrictSeq (NativeScript (AllegraEra c)))

mkRequireAnyOf ∷ StrictSeq (NativeScript (AllegraEra c)) → NativeScript (AllegraEra c)

getRequireAnyOf ∷ NativeScript (AllegraEra c) → Maybe (StrictSeq (NativeScript (AllegraEra c)))

mkRequireMOf ∷ Int → StrictSeq (NativeScript (AllegraEra c)) → NativeScript (AllegraEra c)

getRequireMOf ∷ NativeScript (AllegraEra c) → Maybe (Int, StrictSeq (NativeScript (AllegraEra c)))