Cardano.Ledger.Allegra.Translation

Documentation

shelleyToAllegraAVVMsToDelete ∷ NewEpochState (ShelleyEra c) → UTxO (ShelleyEra c)

Return the subset of UTxO corresponding to Byron-era AVVM addresses, which are to be removed on the Shelley/Allegra boundary. This set will be passed _back_ to the translation functions as the UTxO, allowing these addresses to be removed. This is needed because we cannot do a full scan on the UTxO at this point, since it has been persisted to disk.

Orphan instances

Crypto c ⇒ TranslateEra (AllegraEra c) CertState

Instance details

Associated Types

type TranslationError (AllegraEra c) CertState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → CertState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) CertState) (CertState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) CommitteeState

Instance details

Associated Types

type TranslationError (AllegraEra c) CommitteeState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → CommitteeState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) CommitteeState) (CommitteeState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) DState

Instance details

Associated Types

type TranslationError (AllegraEra c) DState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → DState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) DState) (DState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) PState

Instance details

Associated Types

type TranslationError (AllegraEra c) PState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → PState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) PState) (PState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) VState

Instance details

Associated Types

type TranslationError (AllegraEra c) VState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → VState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) VState) (VState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) PParams

Instance details

Associated Types

type TranslationError (AllegraEra c) PParams

Methods

translateEra ∷ TranslationContext (AllegraEra c) → PParams (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) PParams) (PParams (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) PParamsUpdate

Instance details

Associated Types

type TranslationError (AllegraEra c) PParamsUpdate

Methods

translateEra ∷ TranslationContext (AllegraEra c) → PParamsUpdate (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) PParamsUpdate) (PParamsUpdate (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) UTxO

Instance details

Associated Types

type TranslationError (AllegraEra c) UTxO

Methods

translateEra ∷ TranslationContext (AllegraEra c) → UTxO (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) UTxO) (UTxO (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) FuturePParams

Instance details

Associated Types

type TranslationError (AllegraEra c) FuturePParams

Methods

translateEra ∷ TranslationContext (AllegraEra c) → FuturePParams (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) FuturePParams) (FuturePParams (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) ShelleyGovState

Instance details

Associated Types

type TranslationError (AllegraEra c) ShelleyGovState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → ShelleyGovState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) ShelleyGovState) (ShelleyGovState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) EpochState

Instance details

Associated Types

type TranslationError (AllegraEra c) EpochState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → EpochState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) EpochState) (EpochState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) LedgerState

Instance details

Associated Types

type TranslationError (AllegraEra c) LedgerState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → LedgerState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) LedgerState) (LedgerState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) NewEpochState

Instance details

Associated Types

type TranslationError (AllegraEra c) NewEpochState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → NewEpochState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) NewEpochState) (NewEpochState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) UTxOState

Instance details

Associated Types

type TranslationError (AllegraEra c) UTxOState

Methods

translateEra ∷ TranslationContext (AllegraEra c) → UTxOState (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) UTxOState) (UTxOState (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) ProposedPPUpdates

Instance details

Associated Types

type TranslationError (AllegraEra c) ProposedPPUpdates

Methods

translateEra ∷ TranslationContext (AllegraEra c) → ProposedPPUpdates (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) ProposedPPUpdates) (ProposedPPUpdates (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) Update

Instance details

Associated Types

type TranslationError (AllegraEra c) Update

Methods

translateEra ∷ TranslationContext (AllegraEra c) → Update (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) Update) (Update (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) ShelleyTx

Instance details

Associated Types

type TranslationError (AllegraEra c) ShelleyTx

Methods

translateEra ∷ TranslationContext (AllegraEra c) → ShelleyTx (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) ShelleyTx) (ShelleyTx (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) ShelleyTxOut

Instance details

Associated Types

type TranslationError (AllegraEra c) ShelleyTxOut

Methods

translateEra ∷ TranslationContext (AllegraEra c) → ShelleyTxOut (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) ShelleyTxOut) (ShelleyTxOut (AllegraEra c))

Crypto c ⇒ TranslateEra (AllegraEra c) ShelleyTxWits

Instance details

Associated Types

type TranslationError (AllegraEra c) ShelleyTxWits

Methods

translateEra ∷ TranslationContext (AllegraEra c) → ShelleyTxWits (PreviousEra (AllegraEra c)) → Except (TranslationError (AllegraEra c) ShelleyTxWits) (ShelleyTxWits (AllegraEra c))