Cardano.Ledger.Allegra.TxBody

Documentation

class EraTxBody era ⇒ AllegraEraTxBody era where

Methods

vldtTxBodyL ∷ Lens' (TxBody era) ValidityInterval

data AllegraTxBody e where

Bundled Patterns

pattern AllegraTxBody ∷ (EraTxOut era, EraTxCert era) ⇒ Set (TxIn (EraCrypto era)) → StrictSeq (TxOut era) → StrictSeq (TxCert era) → Withdrawals (EraCrypto era) → Coin → ValidityInterval → StrictMaybe (Update era) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → AllegraTxBody era
A pattern to keep the newtype and the MemoBytes hidden

emptyAllegraTxBodyRaw ∷ Monoid ma ⇒ AllegraTxBodyRaw ma era

data AllegraTxBodyRaw ma era

Constructors

AllegraTxBodyRaw

Fields

atbrInputs ∷ !(Set (TxIn (EraCrypto era)))

atbrOutputs ∷ !(StrictSeq (TxOut era))

atbrCerts ∷ !(StrictSeq (TxCert era))

atbrWithdrawals ∷ !(Withdrawals (EraCrypto era))

atbrTxFee ∷ !Coin

atbrValidityInterval ∷ !ValidityInterval

atbrUpdate ∷ !(StrictMaybe (Update era))

atbrAuxDataHash ∷ !(StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

atbrMint ∷ !ma

data StrictMaybe a

Constructors

SNothing

SJust !a

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore ∷ !(StrictMaybe SlotNo)

invalidHereafter ∷ !(StrictMaybe SlotNo)