Cardano.Ledger.Allegra.TxBody

Documentation

data AllegraTxBody e where

Bundled Patterns

pattern AllegraTxBody ∷ EraTxOut era ⇒ Set (TxIn (EraCrypto era)) → StrictSeq (TxOut era) → StrictSeq (DCert (EraCrypto era)) → Withdrawals (EraCrypto era) → Coin → ValidityInterval → StrictMaybe (Update era) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → AllegraTxBody era
A pattern to keep the newtype and the MemoBytes hidden

class ShelleyEraTxBody era ⇒ AllegraEraTxBody era where

Methods

vldtTxBodyL ∷ Lens' (TxBody era) ValidityInterval

emptyAllegraTxBodyRaw ∷ Monoid ma ⇒ AllegraTxBodyRaw ma era

data AllegraTxBodyRaw ma era

Constructors

AllegraTxBodyRaw

Fields

atbrInputs ∷ !(Set (TxIn (EraCrypto era)))

atbrOutputs ∷ !(StrictSeq (TxOut era))

atbrCerts ∷ !(StrictSeq (DCert (EraCrypto era)))

atbrWithdrawals ∷ !(Withdrawals (EraCrypto era))

atbrTxFee ∷ !Coin

atbrValidityInterval ∷ !ValidityInterval

atbrUpdate ∷ !(StrictMaybe (Update era))

atbrAuxDataHash ∷ !(StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

atbrMint ∷ !ma

data StrictMaybe a

Constructors

SNothing

SJust !a

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore ∷ !(StrictMaybe SlotNo)

invalidHereafter ∷ !(StrictMaybe SlotNo)

Orphan instances

Crypto c ⇒ AllegraEraTxBody (AllegraEra c)

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody (AllegraEra c)) ValidityInterval

Crypto c ⇒ EraTxBody (AllegraEra c)

Instance details

Associated Types

type TxBody (AllegraEra c) = (r ∷ Type)

Methods

mkBasicTxBody ∷ TxBody (AllegraEra c)

inputsTxBodyL ∷ Lens' (TxBody (AllegraEra c)) (Set (TxIn (EraCrypto (AllegraEra c))))

outputsTxBodyL ∷ Lens' (TxBody (AllegraEra c)) (StrictSeq (TxOut (AllegraEra c)))

feeTxBodyL ∷ Lens' (TxBody (AllegraEra c)) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody (AllegraEra c)) (Withdrawals (EraCrypto (AllegraEra c)))

auxDataHashTxBodyL ∷ Lens' (TxBody (AllegraEra c)) (StrictMaybe (AuxiliaryDataHash (EraCrypto (AllegraEra c))))

allInputsTxBodyF ∷ SimpleGetter (TxBody (AllegraEra c)) (Set (TxIn (EraCrypto (AllegraEra c))))

Crypto c ⇒ ShelleyEraTxBody (AllegraEra c)

Instance details

Methods

ttlTxBodyL ∷ Lens' (TxBody (AllegraEra c)) SlotNo

updateTxBodyL ∷ Lens' (TxBody (AllegraEra c)) (StrictMaybe (Update (AllegraEra c)))

updateTxBodyG ∷ SimpleGetter (TxBody (AllegraEra c)) (StrictMaybe (Update (AllegraEra c)))

certsTxBodyL ∷ Lens' (TxBody (AllegraEra c)) (StrictSeq (DCert (EraCrypto (AllegraEra c))))

certsTxBodyG ∷ SimpleGetter (TxBody (AllegraEra c)) (StrictSeq (DCert (EraCrypto (AllegraEra c))))